refactor(cli): move test modules into tests submodule files
Relocate the inline `#[cfg(test)] mod tests` blocks in `progress` and
the binary root into sibling `tests.rs` files (the crate root's submodule
lands at `src/tests.rs`). Production code only in the source files; 22
cli tests unchanged.
